Added feature to remove an expenseList by clicking on trash icon --- no confirmation asked yet

+++ src/main/java/com/wildcodeschool/tricount/controller/ExpenseListController.java
@@ -4,9 +4,13 @@ import com.wildcodeschool.tricount.dto.CreateOrUpdateExpenseListDto;
 import com.wildcodeschool.tricount.entity.ExpenseList;
 import com.wildcodeschool.tricount.service.ExpenseListService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
+
+import java.util.Optional;
 
 @Controller
 public class ExpenseListController {
@@ -44,9 +48,16 @@ public class ExpenseListController {
         return "redirect:/";
     }
 
-    @DeleteMapping("/expenseslist")
-    public String deleteExpensesList(Model model, @RequestParam Integer idList){
-        expenseListService.delete(idList);
+//    @DeleteMapping("/expenseslist")
+    @GetMapping("/expenseslist/delete/{idList}")
+//    public String deleteExpensesList(Model model, @RequestParam Integer idList){
+    public String deleteExpensesList(@PathVariable Integer idList){
+        Optional<ExpenseList> optionalExpenseList = expenseListService.findById(idList);
+        if(optionalExpenseList.isPresent()) {
+            expenseListService.delete(idList);
+        } else {
+                throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Expense List ID not found "+idList);
+        }
         return "redirect:/";
     }
